+++ services/web/server/src/simcore_service_webserver/projects/_crud_handlers_utils.py
@@ -5,7 +5,7 @@ from models_library.rest_pagination_utils import paginate_data
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 from servicelib.rest_constants import RESPONSE_MODEL_POLICY
 
-from . import _permalink_api
+from . import _permalink_service
 from ._crud_api_read import _paralell_update
 from .models import ProjectDict
 
@@ -17,7 +17,7 @@ async def aggregate_data_to_projects_from_request(
 
     update_permalink_per_project = [
         # permalink
-        _permalink_api.aggregate_permalink_in_project(request, project=prj)
+        _permalink_service.aggregate_permalink_in_project(request, project=prj)
         for prj in projects
     ]
