+++ services/web/server/src/simcore_service_webserver/projects/projects_permalink_service.py
@@ -1,5 +1,5 @@
-from ._permalink_api import ProjectPermalink
-from ._permalink_api import register_factory as register_permalink_factory
+from ._permalink_service import ProjectPermalink
+from ._permalink_service import register_factory as register_permalink_factory
 
 __all__: tuple[str, ...] = (
     "ProjectPermalink",
